requirements.txt
Flask
openai
twilio
python-dotenv
requests
